Add UserController test
---

Original commit (chain-contracts): 20ea9f46dd89bf73a05db1780ecd0bb0a4435808

+++ chain-contracts/contracts/controllers/modules/UserNormalPlanetControllable.sol
@@ -77,8 +77,8 @@ contract UserNormalPlanetControllable is PermanenceInterpretable, TimeGettable {
     return records;
   }
 
-  function userNormalPlanetRecordsCountOf(address account) public view returns (uint256) {
-    return _userNormalPlanetPermanence.arrayLength(account);
+  function userNormalPlanetRecordsCountOf(address account) public view returns (uint16) {
+    return uint16(_userNormalPlanetPermanence.arrayLength(account));
   }
 
   function userNormalPlanetRecordOf(address account, uint16 userPlanetId)
